Implement special action.
Instead of the "next" button it is now possible to define actions, like
"click" or "change" in order to expect user to do the action by
themself.
When such action is set, it is not possible to go further with
the "next" button. It is also not possible to come back to previous step
if the previous step use such actions.