Add scroll indicator
When target is out of screen, it now adds an indicator to scroll.
It matches global page scroll and inner element scrolls.
It works with vertical and horizontal scroll.
It computes the correct element to scroll as fast as possible but needs
to recompute on any scroll change.

An SVG icon has been added for this purpose and all arrow SVG have been
changed to be more contrasted with background.