Fix: Export types correctly
The types.d.ts file which is declared in src/ is not integrated in
packages. Consequently, projects using the library does not have the
correct type (they are considered as any).

Creating a symlink doesn't work when creating the package (the file is
still missing). And because hard link doesn't work with git, the npm run
pack script has been modified to create a copy of the file before
creating the package.