fix: avoid starting tutorial when there are no tutorial.
When we change to give no tutorial in props, the computed steps
generates a new empty array. This triggers the watch which starts
tutorial.

To avoids this, we should first check that this.tutorial is defined and
then we use a common reference for empty array to avoid retriggering
reactive properties.