Await that elements are visible
Some elements maybe hidden (like with "display: none"), and selectors
match these elements but it is not ready yet.

So now, it checks that boxes are all ok. If some of elements are hidden
(width, height and x, y are set to 0) then the box is rechecked some
time later.

The timeout is from the step change until all selected elements are
visible.