Rebuild VStep
The workflow of VStep has been rewrite in order to manage all elements
that needed to be fetched in the same way.

Like so it is easier to resynchronize elements when a change occurs in
DOM.

It fix issues, when page are regularly re-redendered. Events are still
bind to the latest elements.

This could be still improved to observe mutation more precisely
(depending on the change we could recompute boxes only).