Allow to change step start
With the "step" property it is now possible to start or restart a
tutorial from a predefined step.

This option should be used with care because it is possible that the
starting step is not in the correct page or state.

It is also possible to move the tutorial directly to this step if the
tutorial is already open.