Fix issue related to logs
log with code 24 prompted a warning because it was not registered in the
error list.

At start, there was a warning because "step" watch is triggered (due to
immediate option), but the step was not found in the step list. This was
because the tutorial was not open yet and so there is nothing inside the
step list.
Now, the watch try to change the currentStep only if the tutorial is
open.