Change behavior of relative target movement.
Before with a nextStep set to "+2" it tried to find the next 2 step
which are available.

This behavior is not easy, it seems more logical and easier to use, to
target the next 2 steps and then to find the next available step.

This target is indeed built to skip directly many steps that we now that
are not relevant, so it is easier to skip them all and do not check if
they are available or not.