Add position stability
Sometime when there are long text in the content, the window may be
displayed differently depending on its position.
So we can be a situation where at position Bottom, there is not enough
space so it moves it to position left, the window is then
re-dimensionned and with this value it considers that the best position
is bottom.
So the window jumped from a position to another.

To avoid this, the window dimension is stored for each position.
So before considering changing position, we now check that we don't
already have a known dimension for this position.